Rework Feed empty view + structure
